requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-cors-headers==3.0.2
django-filter==2.1.0
django-rest-auth==0.9.5
django-tinymce==2.8.0
djangorestframework==3.9.4
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
oauthlib==3.0.1
Pillow==6.0.0
pkg-resources==0.0.0
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.1
